Branches service added

+++ KralInsaat.Common/Entities/ProductEntity.cs
@@ -13,8 +13,8 @@ namespace KralInsaat.Common.Entities
         public BrandEntity? Brand { get; set; }
         public string? ProductName { get; set; }
         public string? ProductDescription { get; set; }
-        public decimal? ProductPrice { get; set; }
-        public decimal? ProductSalePrice { get; set; }
+        public double? ProductPrice { get; set; }
+        public double? ProductSalePrice { get; set; }
         public List<ProductImagesEntity> Images { get; set; } = [];
     } 
 }
